consolidate ADO code with NuGet special casing so they all have ApiVersion.v3
