+++ src/code/RepositorySettings.cs
@@ -811,12 +811,13 @@ namespace Microsoft.PowerShell.PowerShellGet.UtilClasses
             }
             else if (repoUri.AbsoluteUri.EndsWith("index.json", StringComparison.OrdinalIgnoreCase))
             {
-                if (repoUri.AbsoluteUri.Contains("nuget.org"))
-                {
-                    return PSRepositoryInfo.APIVersion.v3;
-                }
+                return PSRepositoryInfo.APIVersion.v3;
+                // if (repoUri.AbsoluteUri.Contains("nuget.org"))
+                // {
+                //     return PSRepositoryInfo.APIVersion.v3;
+                // }
 
-                return PSRepositoryInfo.APIVersion.azureDevOpsFeed;
+                // return PSRepositoryInfo.APIVersion.azureDevOpsFeed;
             }
             else if (repoUri.Scheme.Equals(Uri.UriSchemeFile, StringComparison.OrdinalIgnoreCase))
             {
