+++ src/code/V3ServerAPICalls.cs
@@ -1,3 +1,6 @@
+
+using System.Runtime.CompilerServices;
+using System.Security.Cryptography;
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
@@ -21,17 +24,19 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
         #region Members
         public override PSRepositoryInfo Repository { get; set; }
         private HttpClient _sessionClient { get; set; }
+        private bool _isNuGetRepo { get; set; }
         public FindResponseType v3FindResponseType = FindResponseType.ResponseString;
         private static readonly Hashtable[] emptyHashResponses = new Hashtable[]{};
+        private static readonly string nugetRepoUri = "https://api.nuget.org/v3/index.json";
         private static readonly string resourcesName = "resources";
-        private static readonly string packageBaseAddressName = "PackageBaseAddress/3.0.0";
-        private static readonly string searchQueryServiceName = "SearchQueryService/3.0.0-beta";
-        private static readonly string registrationsBaseUrlName = "RegistrationsBaseUrl/Versioned";
+        private static readonly string itemsName = "items";
+        private static readonly string countName = "count";
+        private static readonly string versionName = "version";
         private static readonly string dataName = "data";
         private static readonly string idName = "id";
-        private static readonly string versionName = "version";
         private static readonly string tagsName = "tags";
-        private static readonly string versionsName = "versions";
+        private static readonly string catalogEntryProperty = "catalogEntry";
+        private static readonly string packageContentProperty = "packageContent";
 
         #endregion
 
@@ -48,6 +53,8 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
             };
 
             _sessionClient = new HttpClient(handler);
+
+            _isNuGetRepo = String.Equals(Repository.Uri.AbsoluteUri, nugetRepoUri, StringComparison.InvariantCultureIgnoreCase);
         }
 
         #endregion
@@ -56,138 +63,41 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
 
         /// <summary>
         /// Find method which allows for searching for all packages from a repository and returns latest version for each.
-        /// Not supported
+        /// Not supported for ADO repository.
         /// </summary>
         public override FindResults FindAll(bool includePrerelease, ResourceType type, out ExceptionDispatchInfo edi)
         {
-            string errMsg = $"Find all is not supported for the repository {Repository.Uri}";
+            string errMsg = $"Find all is not supported for the V3 repository {Repository.Name}";
             edi = ExceptionDispatchInfo.Capture(new InvalidOperationException(errMsg));
 
             return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
         }
 
         /// <summary>
-        /// Find method which allows for searching for packages with tag from a repository and returns latest version for each.
-        /// Examples: Search -Tag "Redis" -Repository PSGallery
-        /// API call: 
-        /// https://azuresearch-ussc.nuget.org/query?q=tags:redis&prerelease=False&semVerLevel=2.0.0
-        /// 
-        /// Azure Artifacts does not support querying on tags, so if support this scenario we need to search on the term and then filter
+        /// Find method which allows for searching for packages with tag(s) from a repository and returns latest version for each.
+        /// This is supported only for the NuGet repository special case, not other V3 repositories.
         /// </summary>
-        public override FindResults FindTags(string[] tags, bool includePrerelease, ResourceType _type, out ExceptionDispatchInfo edi)
+        public override FindResults FindTags(string[] tags, bool includePrerelease, ResourceType type, out ExceptionDispatchInfo edi)
         {
-            List<string> responses = new List<string>();
-            string firstTag = tags[0]; // TODO: better err handle
-
-            Hashtable resourceUrls = FindResourceType(new string[] { searchQueryServiceName, registrationsBaseUrlName }, out edi);
-            if (edi != null)
+            if (_isNuGetRepo)
             {
+                return FindTagsFromNuGetRepo(tags, includePrerelease, out edi);
+            }
+            else
+            {
+                string errMsg = $"Find by Tags is not supported for the V3 repository {Repository.Name}";
+                edi = ExceptionDispatchInfo.Capture(new InvalidOperationException(errMsg));
                 return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
             }
-
-            string searchQueryServiceUrl = resourceUrls[searchQueryServiceName] as string;
-            string registrationsBaseUrl = resourceUrls[registrationsBaseUrlName] as string;
-
-            bool isNuGetRepo = searchQueryServiceUrl.Contains("nuget.org");
-
-            string query = isNuGetRepo ? $"{searchQueryServiceUrl}?q=tags:{firstTag.ToLower()}&prerelease={includePrerelease}&semVerLevel=2.0.0" :
-                          $"{searchQueryServiceUrl}?q={firstTag.ToLower()}&prerelease={includePrerelease}&semVerLevel=2.0.0";
-
-            Console.WriteLine($"query is:  {query}");
-            // 2) call query with tags. (for Azure artifacts) get unique names, see which ones truly match
-            JsonElement[] tagPkgs = GetJsonElementArr(query, dataName, out edi);
-            if (edi != null)
-            {
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-
-            List<string> matchingResponses = new List<string>();
-            string id;
-            string latestVersion;
-            foreach (var pkgId in tagPkgs)
-            { 
-                try
-                {
-                    if (!pkgId.TryGetProperty(idName, out JsonElement idItem) || !pkgId.TryGetProperty(versionName, out JsonElement versionItem))
-                    {
-                        string errMsg = $"FindTag(): Id or Version element could not be found in response.";
-                        edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
-                        return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-                    }
-
-                    id = idItem.ToString();
-                    latestVersion = versionItem.ToString();
-                }
-                catch (Exception e)
-                {
-                    string errMsg = $"FindTag(): Id or Version element could not be parsed from response due to exception {e.Message}.";
-                    edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
-                    return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-                }
-
-                // determine if id matches our wildcard criteria
-                if (isNuGetRepo)
-                {
-                    string response = FindVersionHelper(registrationsBaseUrl, id, latestVersion, out edi);
-                    if (edi != null)
-                    {
-                        return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-                    }
-
-                    matchingResponses.Add(response);
-                }
-                else
-                {
-                    try {
-                        if (!pkgId.TryGetProperty("tags", out JsonElement tagsItem))
-                        {
-                            string errMsg = $"FindTag(): Tag element could not be found in response.";
-                            edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
-                            return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-                        }
-
-                        string[] pkgTags = GetTagsFromJsonElement(tagsItem);
-                        bool isTagMatch = true;
-                        foreach (string rqTag in tags)
-                        {
-                            if (!pkgTags.Contains(rqTag, StringComparer.OrdinalIgnoreCase))
-                            {
-                                isTagMatch = false;
-                                break;
-                            }
-                        }
-
-                        if (isTagMatch)
-                        {
-                            string response = FindVersionHelper(registrationsBaseUrl, id, latestVersion, out edi);
-                            if (edi != null)
-                            {
-                                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-                            }
-
-                            matchingResponses.Add(response);
-                        }
-                    }
-                    catch (Exception e)
-                    {
-                        string errMsg = $"FindTag(): Tags element could not be parsed from response due to exception {e.Message}.";
-                        edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
-                        return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-                    }
-                }
-            }
-
-            FindResults tagsFoundResult = new FindResults(stringResponse: matchingResponses.ToArray(), hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            return tagsFoundResult;
         }
 
         /// <summary>
-        /// This functionality is not supported for V3 protocol server.
         /// Find method which allows for searching for packages with specified Command or DSCResource name.
+        /// Not supported for ADO repository.
         /// </summary>
         public override FindResults FindCommandOrDscResource(string[] tags, bool includePrerelease, bool isSearchingForCommands, out ExceptionDispatchInfo edi)
         {
-            string errMsg = $"Find by CommandName or DSCResource is not supported for {Repository.Name} as it uses the V3 server protocol";
+            string errMsg = $"Find by CommandName or DSCResource is not supported for the V3 server repository {Repository.Name}";
             edi = ExceptionDispatchInfo.Capture(new InvalidOperationException(errMsg));
 
             return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
@@ -197,351 +107,60 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
         /// Find method which allows for searching for single name and returns latest version.
         /// Name: no wildcard support
         /// Examples: Search "Newtonsoft.Json"
-        /// API call: 
-        ///               https://api.nuget.org/v3/registration5-gz-semver2/nuget.server/index.json
-        ///               https://msazure.pkgs.visualstudio.com/One/_packaging/testfeed/nuget/v3/registrations2-semver2/newtonsoft.json/index.json
-        ///               https://msazure.pkgs.visualstudio.com/999aa88e-7ed7-41b2-9d77-5bc261222004/_packaging/0d5429e2-c871-4347-bdc9-d1cbbac5eb3b/nuget/v3/registrations2-semver2/newtonsoft.json/index.json
-        /// The RegistrationBaseUrl that we're using is "RegistrationBaseUrl/Versioned"
-        /// This type points to the url to use (ex above)
-        /// Implementation Note: Need to filter further for latest version (prerelease or non-prerelease dependening on user preference)
+        /// We use the latest RegistrationBaseUrl version resource we can find and check if contains an entry with the package name.
         /// </summary>
         public override FindResults FindName(string packageName, bool includePrerelease, ResourceType type, out ExceptionDispatchInfo edi)
         {
-            Hashtable resourceUrls = FindResourceType(new string[] { packageBaseAddressName, registrationsBaseUrlName }, out edi);
-            if (edi != null)
-            {
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-
-            string packageBaseAddressUrl = resourceUrls[packageBaseAddressName] as string;
-            string registrationsBaseUrl = resourceUrls[registrationsBaseUrlName] as string;
-
-            bool isNuGetRepo = packageBaseAddressUrl.Contains("v3-flatcontainer");
-            JsonElement[] pkgVersionsArr = GetPackageVersions(packageBaseAddressUrl, packageName, isNuGetRepo, out edi);
-            if (edi != null)
-            {
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-
-            string response = string.Empty;
-            foreach (JsonElement version in pkgVersionsArr)
-            {
-                // parse as NuGetVersion
-                if (NuGetVersion.TryParse(version.ToString(), out NuGetVersion nugetVersion))
-                {
-                    /* 
-                     * pkgVersion == !prerelease   &&   includePrerelease == true   -->   keep pkg   
-                     * pkgVersion == !prerelease   &&   includePrerelease == false  -->   keep pkg 
-                     * pkgVersion == prerelease    &&   includePrerelease == true   -->   keep pkg   
-                     * pkgVersion == prerelease    &&   includePrerelease == false  -->   throw away pkg 
-                     */
-                    if (!nugetVersion.IsPrerelease || includePrerelease)
-                    {
-                        response = FindVersionHelper(registrationsBaseUrl, packageName, version.ToString(), out edi);
-                        if (edi != null)
-                        {
-                            return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-                        }
-
-                        break;
-                    }
-                }
-            }
-
-            if (String.IsNullOrEmpty(response))
-            {
-                edi = ExceptionDispatchInfo.Capture(new InvalidOrEmptyResponse($"FindName() with {packageName} returned empty response."));
-            }
-
-            return new FindResults(stringResponse: new string[] { response }, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            return FindNameHelper(packageName, tags: Utils.EmptyStrArray, includePrerelease, type, out edi);
         }
 
         /// <summary>
-        /// Find method which allows for searching for single name and tag and returns latest version.
+        /// Find method which allows for searching for single name and specified tag(s) and returns latest version.
         /// Name: no wildcard support
-        /// Examples: Search "Newtonsoft.Json" - Tag "json"
+        /// Examples: Search "Newtonsoft.Json" -Tag "json"
+        /// We use the latest RegistrationBaseUrl version resource we can find and check if contains an entry with the package name.
         /// </summary>
         public override FindResults FindNameWithTag(string packageName, string[] tags, bool includePrerelease, ResourceType type, out ExceptionDispatchInfo edi)
         {
-            Hashtable resourceUrls = FindResourceType(new string[] { packageBaseAddressName, registrationsBaseUrlName }, out edi);
-            if (edi != null)
-            {
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-
-            string packageBaseAddressUrl = resourceUrls[packageBaseAddressName] as string;
-            string registrationsBaseUrl = resourceUrls[registrationsBaseUrlName] as string;
-
-            bool isNuGetRepo = packageBaseAddressUrl.Contains("v3-flatcontainer");
-            JsonElement[] pkgVersionsArr = GetPackageVersions(packageBaseAddressUrl, packageName, isNuGetRepo, out edi);
-            if (edi != null)
-            {
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-
-            string response = string.Empty;
-            foreach (JsonElement version in pkgVersionsArr)
-            {
-                // parse as NuGetVersion
-                if (NuGetVersion.TryParse(version.ToString(), out NuGetVersion nugetVersion))
-                {
-                    /* 
-                     * pkgVersion == !prerelease   &&   includePrerelease == true   -->   keep pkg   
-                     * pkgVersion == !prerelease   &&   includePrerelease == false  -->   keep pkg 
-                     * pkgVersion == prerelease    &&   includePrerelease == true   -->   keep pkg   
-                     * pkgVersion == prerelease    &&   includePrerelease == false  -->   throw away pkg 
-                     */
-                    if (!nugetVersion.IsPrerelease || includePrerelease)
-                    {
-                        response = FindVersionHelper(registrationsBaseUrl, packageName, version.ToString(), out edi);
-                        if (edi != null)
-                        {
-                            return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-                        }
-
-                        bool isTagMatch = DetermineTagsPresent(response: response, tags: tags, out edi);
-
-                        if (!isTagMatch)
-                        {
-                            if (edi == null)
-                            {
-                                string errMsg = $"FindNameWithTag(): Tags required were not found in package {packageName} {version.ToString()}.";
-                                edi = ExceptionDispatchInfo.Capture(new SpecifiedTagsNotFoundException(errMsg));
-                            }
-
-                            return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-                        }
-
-                        break;
-                    }
-                }
-            }
-
-            if (String.IsNullOrEmpty(response))
-            {
-                edi = ExceptionDispatchInfo.Capture(new InvalidOrEmptyResponse($"FindNameWithTag() with {packageName} and tags {String.Join(",", tags)} returned empty response."));
-            }
-
-            return new FindResults(stringResponse: new string[] { response }, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            return FindNameHelper(packageName, tags, includePrerelease, type, out edi);
         }
 
         /// <summary>
         /// Find method which allows for searching for single name with wildcards and returns latest version.
-        /// Name: supports wildcards
-        /// Examples: Search "Nuget.Server*"
-        /// API call: 
-        /// - No prerelease: https://api-v2v3search-0.nuget.org/autocomplete?q=storage&prerelease=false
-        /// - Prerelease:  https://api-v2v3search-0.nuget.org/autocomplete?q=storage&prerelease=true  
-        /// 
-        /// https://msazure.pkgs.visualstudio.com/b32aa71e-8ed2-41b2-9d77-5bc261222004/_packaging/0d5429e2-c871-4347-bdc9-d1cbbac5eb3b/nuget/v3/query2?q=Newtonsoft&prerelease=false&semVerLevel=2.0.0
-        ///         
-        ///        Note:  response only returns names
-        ///        
-        ///        Make another query to get the latest version of each package  (ie call "FindVersionGlobbing")
+        /// This is supported only for the NuGet repository special case, not other V3 repositories.
         /// </summary>
         public override FindResults FindNameGlobbing(string packageName, bool includePrerelease, ResourceType type, out ExceptionDispatchInfo edi)
         {
-            var names = packageName.Split(new char[] { '*' }, StringSplitOptions.RemoveEmptyEntries);
-            string querySearchTerm;
-
-            if (names.Length == 0)
+            if (_isNuGetRepo)
             {
-                edi = ExceptionDispatchInfo.Capture(new ArgumentException("-Name '*' for V3 server protocol repositories is not supported"));
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-            if (names.Length == 1)
-            {
-                // packageName: *get*       -> q: get
-                // packageName: PowerShell* -> q: PowerShell
-                // packageName: *ShellGet   -> q: ShellGet
-                querySearchTerm = names[0];
+                return FindNameGlobbingFromNuGetRepo(packageName, tags: Utils.EmptyStrArray, includePrerelease, out edi);
             }
             else
             {
-                // *pow*get*
-                // pow*get -> only support this (V2)
-                // pow*get*
-                // *pow*get
+                string errMsg = $"Find with Name containing wildcards is not supported for the V3 server repository {Repository.Name}";
+                edi = ExceptionDispatchInfo.Capture(new InvalidOperationException(errMsg));
 
-                edi = ExceptionDispatchInfo.Capture(new ArgumentException("-Name with wildcards is only supported for scenarios similar to the following examples: PowerShell*, *ShellGet, *Shell*."));
                 return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
             }
-
-            // https://msazure.pkgs.visualstudio.com/.../_packaging/.../nuget/v3/query2 (no support for * in search term, but matches like NuGet)
-            // https://azuresearch-usnc.nuget.org/query?q=Newtonsoft&prerelease=false&semVerLevel=1.0.0 (NuGet) (supports * at end of searchterm q but equivalent to q = text w/o *)
-            Hashtable resourceUrls = FindResourceType(new string[] { searchQueryServiceName, registrationsBaseUrlName }, out edi);
-            if (edi != null)
-            {
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-
-            string searchQueryServiceUrl = resourceUrls[searchQueryServiceName] as string;
-            string registrationsBaseUrl = resourceUrls[registrationsBaseUrlName] as string;
-
-            string query = $"{searchQueryServiceUrl}?q={querySearchTerm}&prerelease={includePrerelease}&semVerLevel=2.0.0";
-
-            // 2) call query with search term, get unique names, see which ones truly match
-            JsonElement[] matchingPkgIds = GetJsonElementArr(query, dataName, out edi);
-            if (edi != null)
-            {
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-
-            List<string> matchingResponses = new List<string>();
-            foreach (var pkgId in matchingPkgIds)
-            {
-                string id = string.Empty;
-                string latestVersion = string.Empty;
-
-                try
-                {
-                    if (!pkgId.TryGetProperty(idName, out JsonElement idItem) || ! pkgId.TryGetProperty(versionName, out JsonElement versionItem))
-                    {
-                        string errMsg = $"FindNameGlobbing(): Name or Version element could not be found in response.";
-                        edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
-                        return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-                    }
-
-                    id = idItem.ToString();
-                    latestVersion = versionItem.ToString();
-                }
-                catch (Exception e)
-                {
-                    string errMsg = $"FindNameGlobbing(): Name or Version element could not be parsed from response due to exception {e.Message}.";
-                    edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
-                    break;
-                }
-
-                // determine if id matches our wildcard criteria
-                if ((packageName.StartsWith("*") && packageName.EndsWith("*") && id.ToLower().Contains(querySearchTerm.ToLower())) ||
-                    (packageName.EndsWith("*") && id.StartsWith(querySearchTerm, StringComparison.OrdinalIgnoreCase)) ||
-                    (packageName.StartsWith("*") && id.EndsWith(querySearchTerm, StringComparison.OrdinalIgnoreCase)))
-                {
-                    string response = FindVersionHelper(registrationsBaseUrl, id, latestVersion, out edi);
-
-                    if (edi != null)
-                    {
-                        return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-                    }
-
-                    matchingResponses.Add(response);
-                }
-            }
-
-            return new FindResults(stringResponse: matchingResponses.ToArray(), hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
         }
 
         /// <summary>
         /// Find method which allows for searching for single name with wildcards and tag and returns latest version.
-        /// Name: supports wildcards
-        /// Examples: Search "Nuget.Server*" -Tag "nuget"
+        /// This is supported only for the NuGet repository special case, not other V3 repositories.
         /// </summary>
         public override FindResults FindNameGlobbingWithTag(string packageName, string[] tags, bool includePrerelease, ResourceType type, out ExceptionDispatchInfo edi)
         {
-            var names = packageName.Split(new char[] { '*' }, StringSplitOptions.RemoveEmptyEntries);
-            string querySearchTerm;
-
-            if (names.Length == 0)
+            if (_isNuGetRepo)
             {
-                edi = ExceptionDispatchInfo.Capture(new ArgumentException("-Name '*' for V3 server protocol repositories is not supported"));
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-            if (names.Length == 1)
-            {
-                // packageName: *get*       -> q: get
-                // packageName: PowerShell* -> q: PowerShell
-                // packageName: *ShellGet   -> q: ShellGet
-                querySearchTerm = names[0];
+                return FindNameGlobbingFromNuGetRepo(packageName, tags, includePrerelease, out edi);
             }
             else
             {
-                // *pow*get*
-                // pow*get -> only support this (V2)
-                // pow*get*
-                // *pow*get
+                string errMsg = $"Find with Name containing wildcards is not supported for the V3 server repository {Repository.Name}";
+                edi = ExceptionDispatchInfo.Capture(new InvalidOperationException(errMsg));
 
-                edi = ExceptionDispatchInfo.Capture(new ArgumentException("-Name with wildcards is only supported for scenarios similar to the following examples: PowerShell*, *ShellGet, *Shell*."));
                 return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
             }
-
-            // https://msazure.pkgs.visualstudio.com/.../_packaging/.../nuget/v3/query2 (no support for * in search term, but matches like NuGet)
-            // https://azuresearch-usnc.nuget.org/query?q=Newtonsoft&prerelease=false&semVerLevel=1.0.0 (NuGet) (supports * at end of searchterm q but equivalent to q = text w/o *)
-            Hashtable resourceUrls = FindResourceType(new string[] { searchQueryServiceName, registrationsBaseUrlName }, out edi);
-            if (edi != null)
-            {
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-
-            string searchQueryServiceUrl = resourceUrls[searchQueryServiceName] as string;
-            string registrationsBaseUrl = resourceUrls[registrationsBaseUrlName] as string;
-
-            string query = $"{searchQueryServiceUrl}?q={querySearchTerm}&prerelease={includePrerelease}&semVerLevel=2.0.0";
-
-            // 2) call query with search term, get unique names, see which ones truly match
-            JsonElement[] matchingPkgIds = GetJsonElementArr(query, dataName, out edi);
-            if (edi != null)
-            {
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-
-            List<string> matchingResponses = new List<string>();
-            foreach (var pkgId in matchingPkgIds)
-            {
-                string id = string.Empty;
-                string latestVersion = string.Empty;
-                string[] pkgTags = Utils.EmptyStrArray;
-
-                try
-                {
-                    if (!pkgId.TryGetProperty(idName, out JsonElement idItem) || !pkgId.TryGetProperty(versionName, out JsonElement versionItem))
-                    {
-                        string errMsg = $"FindNameGlobbing(): Name or Version element could not be found in response.";
-                        edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
-                        return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-                    }
-
-                    if (!pkgId.TryGetProperty(tagsName, out JsonElement tagsItem))
-                    {
-                        string errMsg = $"FindNameGlobbing(): Tags element could not be found in response.";
-                        edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
-                        return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-                    }
-
-                    id = idItem.ToString();
-                    latestVersion = versionItem.ToString();
-
-                    pkgTags = GetTagsFromJsonElement(tagsElement: tagsItem);
-                }
-                catch (Exception e)
-                {
-                    string errMsg = $"FindNameGlobbingWithTag(): Name or Version or Tags element could not be parsed from response due to exception {e.Message}.";
-                    edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
-                    break;
-                }
-
-                // determine if id matches our wildcard criteria
-                if ((packageName.StartsWith("*") && packageName.EndsWith("*") && id.ToLower().Contains(querySearchTerm.ToLower())) ||
-                    (packageName.EndsWith("*") && id.StartsWith(querySearchTerm, StringComparison.OrdinalIgnoreCase)) ||
-                    (packageName.StartsWith("*") && id.EndsWith(querySearchTerm, StringComparison.OrdinalIgnoreCase)))
-                {
-                    bool isTagMatch = DeterminePkgTagsSatisfyRequiredTags(pkgTags: pkgTags, requiredTags: tags);
-                    if (!isTagMatch)
-                    {
-                        continue;
-                    }
-
-                    string response = FindVersionHelper(registrationsBaseUrl, id, latestVersion, out edi);
-
-                    if (edi != null)
-                    {
-                        continue;
-                    }
-
-                    matchingResponses.Add(response);
-                }
-            }
-
-            return new FindResults(stringResponse: matchingResponses.ToArray(), hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
         }
 
         /// <summary>
@@ -550,62 +169,59 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
         /// Version: supports wildcards
         /// Examples: Search "NuGet.Server.Core" "[1.0.0.0, 5.0.0.0]"
         ///           Search "NuGet.Server.Core" "3.*"
-        /// API Call: 
-        ///           then, find all versions for a pkg
-        ///           for nuget:
-        ///               this contains all pkg version info: https://api.nuget.org/v3/registration5-gz-semver2/nuget.server/index.json
-        ///               However, we will use the flattened version list: https://api.nuget.org/v3-flatcontainer/newtonsoft.json/index.json
-        ///           for Azure Artifacts:
-        ///               https://msazure.pkgs.visualstudio.com/b32aa71e-8ed2-41b2-9d77-5bc261222004/_packaging/0d5429e2-c871-4347-bdc9-d1cbbac5eb3b/nuget/v3/flat2/newtonsoft.json/index.json
-        ///            (azure artifacts)
-        ///            
-        ///             Note:  very different responses for nuget vs azure artifacts
-        ///            
-        ///            After we figure out what version we want, call "FindVersion" (or some helper method)
-        /// need to filter client side
-        /// Implementation note: Returns all versions, including prerelease ones. Later (in the API client side) we'll do filtering on the versions to satisfy what user provided.
+        /// We use the latest RegistrationBaseUrl version resource we can find and check if contains an entry with the package name, then get all versions and match to satisfying versions.
         /// </summary>
         public override FindResults FindVersionGlobbing(string packageName, VersionRange versionRange, bool includePrerelease, ResourceType type, bool getOnlyLatest, out ExceptionDispatchInfo edi)
         {
-            Hashtable resourceUrls = FindResourceType(new string[] { packageBaseAddressName, registrationsBaseUrlName }, out edi);
+            // Get all the resources for the service index
+            Dictionary<string, string> resources = GetResourcesFromServiceIndex(out edi);
             if (edi != null)
             {
                 return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
             }
 
-            string packageBaseAddressUrl = resourceUrls[packageBaseAddressName] as string;
-            string registrationsBaseUrl = resourceUrls[registrationsBaseUrlName] as string;
-
-            bool isNuGetRepo = packageBaseAddressUrl.Contains("v3-flatcontainer");
-            JsonElement[] pkgVersionsArr = GetPackageVersions(packageBaseAddressUrl, packageName, isNuGetRepo, out edi);
+            string registrationsBaseUrl = FindRegistrationsBaseUrlForAllV3(resources, out edi);
             if (edi != null)
             {
                 return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
             }
 
-            List<string> responses = new List<string>();
-            foreach (var version in pkgVersionsArr) {
-                if (NuGetVersion.TryParse(version.ToString(), out NuGetVersion nugetVersion) && versionRange.Satisfies(nugetVersion))
+            string[] versionedResponses = GetVersionedResponses(registrationsBaseUrl, packageName, catalogEntryProperty, out edi);
+            if (edi != null)
+            {
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            List<string> satisfyingVersions = new List<string>();
+            foreach (string response in versionedResponses)
+            {
+                JsonElement pkgVersionElement;
+                try
                 {
-                    /* 
-                     * pkgVersion == !prerelease   &&   includePrerelease == true   -->   keep pkg   
-                     * pkgVersion == !prerelease   &&   includePrerelease == false  -->   keep pkg 
-                     * pkgVersion == prerelease    &&   includePrerelease == true   -->   keep pkg   
-                     * pkgVersion == prerelease    &&   includePrerelease == false  -->   throw away pkg 
-                     */
-                    if (!nugetVersion.IsPrerelease || includePrerelease) {
-                        string response = FindVersionHelper(registrationsBaseUrl, packageName, version.ToString(), out edi);
-                        if (edi != null)
-                        {
-                            return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-                        }
+                    JsonDocument pkgVersionEntry = JsonDocument.Parse(response);
+                    JsonElement rootDom = pkgVersionEntry.RootElement;
+                    if (!rootDom.TryGetProperty(versionName, out pkgVersionElement))
+                    {
+                        edi = ExceptionDispatchInfo.Capture(new InvalidOrEmptyResponse($"Response does not contain 'version' element."));
+                        return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+                    }
+                }
+                catch (Exception e)
+                {
+                    edi = ExceptionDispatchInfo.Capture(e);
+                    return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+                }
 
-                        responses.Add(response);
+                if (NuGetVersion.TryParse(pkgVersionElement.ToString(), out NuGetVersion pkgVersion) && versionRange.Satisfies(pkgVersion))
+                {
+                    if (!pkgVersion.IsPrerelease || includePrerelease)
+                    {
+                        satisfyingVersions.Add(response);
                     }
                 }
             }
 
-            return new FindResults(stringResponse: responses.ToArray(), hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            return new FindResults(stringResponse: satisfyingVersions.ToArray(), hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
         }
 
         /// <summary>
@@ -613,100 +229,23 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
         /// Name: no wildcard support
         /// Version: no wildcard support
         /// Examples: Search "NuGet.Server.Core" "3.0.0-beta"
-        /// API call: 
-        ///     first find the RegistrationBaseUrl
-        ///     https://api.nuget.org/v3/registration5-gz-semver2/nuget.server/index.json
-        ///     
-        ///     https://msazure.pkgs.visualstudio.com/One/_packaging/testfeed/nuget/v3/registrations2-semver2/newtonsoft.json/index.json
-        ///     https://msazure.pkgs.visualstudio.com/999aa88e-7ed7-41b2-9d77-5bc261222004/_packaging/0d5429e2-c871-4347-bdc9-d1cbbac5eb3b/nuget/v3/registrations2-semver2/newtonsoft.json/index.json
-        ///         The RegistrationBaseUrl that we're using is "RegistrationBaseUrl/Versioned"
-        ///         This type points to the url to use (ex above)
-        ///         
-        ///     then we can make a call for the specific version  
-        ///     https://api.nuget.org/v3/registration5-gz-semver2/nuget.server.core/3.0.0-beta
-        ///     (alternative url for nuget gallery):  https://api.nuget.org/v3/registration5-gz-semver2/nuget.server.core/index.json#page/3.0.0-beta/3.0.0-beta
-        ///     https://msazure.pkgs.visualstudio.com/b32aa71e-8ed2-41b2-9d77-5bc261222004/_packaging/0d5429e2-c871-4347-bdc9-d1cbbac5eb3b/nuget/v3/registrations2/newtonsoft.json/13.0.2.json 
-        ///     
+        /// We use the latest RegistrationBaseUrl version resource we can find and check if contains an entry with the package name, then match to the specified version.
         /// </summary>
         public override FindResults FindVersion(string packageName, string version, ResourceType type, out ExceptionDispatchInfo edi)
         {
-            Hashtable resourceUrls = FindResourceType(new string[] { registrationsBaseUrlName }, out edi);
-            if (edi != null)
-            {
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-
-            string registrationsBaseUrl = resourceUrls[registrationsBaseUrlName] as string;
-
-            string response = FindVersionHelper(registrationsBaseUrl, packageName, version, out edi);
-            if (edi != null)
-            {
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-
-            if (String.IsNullOrEmpty(response))
-            {
-                edi = ExceptionDispatchInfo.Capture(new InvalidOrEmptyResponse($"FindVersion() with {packageName} and version {version} returned empty response."));
-            }
-
-            return new FindResults(stringResponse: new string[] { response }, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            return FindVersionHelper(packageName, version, tags: Utils.EmptyStrArray, type, out edi);
         }
 
         /// <summary>
-        /// Find method which allows for searching for single name with specific version and tag.
+        /// Find method which allows for searching for single name with specific version and tag(s).
         /// Name: no wildcard support
         /// Version: no wildcard support
-        /// Examples: Search "NuGet.Server.Core" -Version "3.0.0-beta" -Tag "nuget"
-        /// API call: 
-        ///     first find the RegistrationBaseUrl
-        ///     https://api.nuget.org/v3/registration5-gz-semver2/nuget.server/index.json
-        ///     
-        ///     https://msazure.pkgs.visualstudio.com/One/_packaging/testfeed/nuget/v3/registrations2-semver2/newtonsoft.json/index.json
-        ///     https://msazure.pkgs.visualstudio.com/999aa88e-7ed7-41b2-9d77-5bc261222004/_packaging/0d5429e2-c871-4347-bdc9-d1cbbac5eb3b/nuget/v3/registrations2-semver2/newtonsoft.json/index.json
-        ///         The RegistrationBaseUrl that we're using is "RegistrationBaseUrl/Versioned"
-        ///         This type points to the url to use (ex above)
-        ///         
-        ///     then we can make a call for the specific version  
-        ///     https://api.nuget.org/v3/registration5-gz-semver2/nuget.server.core/3.0.0-beta
-        ///     (alternative url for nuget gallery):  https://api.nuget.org/v3/registration5-gz-semver2/nuget.server.core/index.json#page/3.0.0-beta/3.0.0-beta
-        ///     https://msazure.pkgs.visualstudio.com/b32aa71e-8ed2-41b2-9d77-5bc261222004/_packaging/0d5429e2-c871-4347-bdc9-d1cbbac5eb3b/nuget/v3/registrations2/newtonsoft.json/13.0.2.json 
-        ///     
-        /// </summary>        
+        /// Examples: Search "NuGet.Server.Core" "3.0.0-beta" -Tag "core"
+        /// We use the latest RegistrationBaseUrl version resource we can find and check if contains an entry with the package name, then match to the specified version.
+        /// </summary>     
         public override FindResults FindVersionWithTag(string packageName, string version, string[] tags, ResourceType type, out ExceptionDispatchInfo edi)
         {
-            Hashtable resourceUrls = FindResourceType(new string[] { registrationsBaseUrlName }, out edi);
-            if (edi != null)
-            {
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-
-            string registrationsBaseUrl = resourceUrls[registrationsBaseUrlName] as string;
-
-            string response = FindVersionHelper(registrationsBaseUrl, packageName, version, out edi);
-            if (edi != null)
-            {
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-
-            bool isTagMatch = DetermineTagsPresent(response: response, tags: tags, out edi);
-
-            if (!isTagMatch)
-            {
-                if (edi == null)
-                {
-                    string errMsg = $"FindVersionWithTag(): Tags required were not found in package {packageName} {version.ToString()}.";
-                    edi = ExceptionDispatchInfo.Capture(new SpecifiedTagsNotFoundException(errMsg));
-                }
-
-                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
-            }
-
-            if (String.IsNullOrEmpty(response))
-            {
-                edi = ExceptionDispatchInfo.Capture(new InvalidOrEmptyResponse($"FindVersion() with {packageName}, tags {String.Join(", ", tags)} and version {version} returned empty response."));
-            }
-
-            return new FindResults(stringResponse: new string[] { response }, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            return FindVersionHelper(packageName, version, tags: tags, type, out edi);
         }
 
         /**  INSTALL APIS **/
@@ -714,85 +253,119 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
         /// <summary>
         /// Installs specific package.
         /// Name: no wildcard support.
-        /// Examples: Install "PowerShellGet"
-        /// Implementation Note:   if not prerelease: https://www.powershellgallery.com/api/v2/package/powershellget (Returns latest stable)
-        ///                        if prerelease, the calling method should first call IFindPSResource.FindName(), 
-        ///                             then find the exact version to install, then call into install version
+        /// Examples: Install "Newtonsoft.json"
         /// </summary>
         public override Stream InstallName(string packageName, bool includePrerelease, out ExceptionDispatchInfo edi)
         {
-            Hashtable resourceUrls = FindResourceType(new string[] { packageBaseAddressName }, out edi);
+            Stream pkgStream = null;
+            // Get all the resources for the service index
+            Dictionary<string, string> resources = GetResourcesFromServiceIndex(out edi);
             if (edi != null)
             {
+                return pkgStream;
+            }
+
+            string registrationsBaseUrl = FindRegistrationsBaseUrlForAllV3(resources, out edi);
+            if (edi != null)
+            {
+                return pkgStream;
+            }
+
+            string[] versionedResponses = GetVersionedResponses(registrationsBaseUrl, packageName, packageContentProperty, out edi);
+            if (edi != null)
+            {
+                return pkgStream;
+            }
+
+            if (versionedResponses.Length == 0)
+            {
+                string errorMsg = $"Package with Name {packageName} could not be found in repository {Repository.Name}";
+                edi = ExceptionDispatchInfo.Capture(new Exception(errorMsg));
+                return pkgStream;
+            }
+
+            // get the Url for the latest version only
+            string pkgContentUrl = versionedResponses[0];
+            var content = HttpRequestCallForContent(pkgContentUrl, out edi);
+            if (edi != null)
+            {
+                string errorMsg = $"Package with Name {packageName} could not be found in repository {Repository.Name}";
+                edi = ExceptionDispatchInfo.Capture(new Exception(errorMsg));
                 return null;
             }
 
-            string packageBaseAddressUrl = resourceUrls[packageBaseAddressName] as string;
-
-            bool isNuGetRepo = packageBaseAddressUrl.Contains("v3-flatcontainer");
-
-            JsonElement[] pkgVersionsArr = GetPackageVersions(packageBaseAddressUrl, packageName, isNuGetRepo, out edi);
-            if (edi != null)
-            {
-                return null;
-            }
-
-            foreach (JsonElement version in pkgVersionsArr)
-            {
-                if (NuGetVersion.TryParse(version.ToString(), out NuGetVersion nugetVersion))
-                {
-                    /* 
-                     * pkgVersion == !prerelease   &&   includePrerelease == true   -->   keep pkg   
-                     * pkgVersion == !prerelease   &&   includePrerelease == false  -->   keep pkg 
-                     * pkgVersion == prerelease    &&   includePrerelease == true   -->   keep pkg   
-                     * pkgVersion == prerelease    &&   includePrerelease == false  -->   throw away pkg 
-                     */
-                    if (!nugetVersion.IsPrerelease || includePrerelease)
-                    {
-                        var responseStream = InstallVersion(packageName, version.ToString(), out edi);
-                        if (edi != null)
-                        {
-                            return null;
-                        }
-
-                        return responseStream;
-                    }
-                }
-            }
-
-            return null;
+            pkgStream = content.ReadAsStreamAsync().Result;
+            return pkgStream;
         }
 
         /// <summary>
         /// Installs package with specific name and version.
         /// Name: no wildcard support.
         /// Version: no wildcard support.
-        /// Examples: Install "PowerShellGet" -Version "3.0.0.0"
-        ///           Install "PowerShellGet" -Version "3.0.0-beta16"
-        ///           
-        ///  https://api.nuget.org/v3-flatcontainer/newtonsoft.json/9.0.1/newtonsoft.json.9.0.1.nupkg
-        /// API Call: 
+        /// Examples: Install "Newtonsoft.json" -Version "1.0.0.0"
+        ///           Install "Newtonsoft.json" -Version "2.5.0-beta"
         /// </summary>    
         public override Stream InstallVersion(string packageName, string version, out ExceptionDispatchInfo edi)
         {
-            Hashtable resourceUrls = FindResourceType(new string[] { packageBaseAddressName }, out edi);
+            Stream pkgStream = null;
+            if (!NuGetVersion.TryParse(version, out NuGetVersion requiredVersion))
+            {
+                edi = ExceptionDispatchInfo.Capture(new ArgumentException($"Version {version} to be installed is not a valid NuGet version."));
+                return pkgStream;
+            }
+
+            // Get all the resources for the service index
+            Dictionary<string, string> resources = GetResourcesFromServiceIndex(out edi);
+            if (edi != null)
+            {
+                return pkgStream;
+            }
+
+            string registrationsBaseUrl = FindRegistrationsBaseUrlForAllV3(resources, out edi);
+            if (edi != null)
+            {
+                return pkgStream;
+            }
+
+            string[] versionedResponses = GetVersionedResponses(registrationsBaseUrl, packageName, packageContentProperty, out edi);
+            if (edi != null)
+            {
+                return pkgStream;
+            }
+
+            if (versionedResponses.Length == 0)
+            {
+                string errorMsg = $"Package with Name {packageName} and Version {version} could not be found in repository {Repository.Name}";
+                edi = ExceptionDispatchInfo.Capture(new Exception(errorMsg));
+                return null;
+            }
+
+            string pkgContentUrl = String.Empty;
+            foreach (string response in versionedResponses)
+            {
+                // Response will be "packageContent" element value that looks like: "{packageBaseAddress}/{packageName}/{normalizedVersion}/{packageName}.{normalizedVersion}.nupkg"
+                if (response.Contains(requiredVersion.ToNormalizedString()))
+                {
+                    pkgContentUrl = response;
+                    break;
+                }
+            }
+
+            if (String.IsNullOrEmpty(pkgContentUrl))
+            {
+                string errorMsg = $"Package with Name {packageName} and Version {version} could not be found in repository {Repository.Name}";
+                edi = ExceptionDispatchInfo.Capture(new Exception(errorMsg));
+                return null;
+            }
+
+            var content = HttpRequestCallForContent(pkgContentUrl, out edi);
             if (edi != null)
             {
                 return null;
             }
 
-            string packageBaseAddressUrl = resourceUrls[packageBaseAddressName] as string;
-
-            string pkgName = packageName.ToLower();
-            string installPkgUrl = $"{packageBaseAddressUrl}{pkgName}/{version}/{pkgName}.{version}.nupkg";
-
-            var content = HttpRequestCallForContent(installPkgUrl, out edi);
-            if (edi != null)
-            {
-                return null;
-            }
-
-            return content.ReadAsStreamAsync().Result;
+            pkgStream = content.ReadAsStreamAsync().Result;
+            return pkgStream;
         }
 
         #endregion
@@ -800,9 +373,321 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
         #region Private Methods
 
         /// <summary>
+        /// Helper method called by FindNameGlobbing() and FindNameGlobbingWithTag() for special case where repository is NuGet.org repository.
+        /// </summary>
+        private FindResults FindNameGlobbingFromNuGetRepo(string packageName, string[] tags, bool includePrerelease, out ExceptionDispatchInfo edi)
+        {
+            var names = packageName.Split(new char[] { '*' }, StringSplitOptions.RemoveEmptyEntries);
+            string querySearchTerm;
+
+            if (names.Length == 0)
+            {
+                edi = ExceptionDispatchInfo.Capture(new ArgumentException("-Name '*' for V3 server protocol repositories is not supported"));
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+            if (names.Length == 1)
+            {
+                // packageName: *get*       -> q: get
+                // packageName: PowerShell* -> q: PowerShell
+                // packageName: *ShellGet   -> q: ShellGet
+                querySearchTerm = names[0];
+            }
+            else
+            {
+                // *pow*get*
+                // pow*get -> only support this (V2)
+                // pow*get*
+                // *pow*get
+
+                edi = ExceptionDispatchInfo.Capture(new ArgumentException("-Name with wildcards is only supported for scenarios similar to the following examples: PowerShell*, *ShellGet, *Shell*."));
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            Dictionary<string, string> resources = GetResourcesFromServiceIndex(out edi);
+            if (edi != null)
+            {
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            string searchQueryServiceUrl = FindSearchQueryService(resources, out edi);
+            if (edi != null)
+            {
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            string query = $"{searchQueryServiceUrl}?q={querySearchTerm}&prerelease={includePrerelease}&semVerLevel=2.0.0";
+
+            JsonElement[] matchingPkgIds = GetJsonElementArr(query, dataName, out edi);
+            if (edi != null)
+            {
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            List<string> matchingResponses = new List<string>();
+            foreach (var pkgEntry in matchingPkgIds)
+            {
+                string id = string.Empty;
+                string latestVersion = string.Empty;
+                string[] pkgTags = Utils.EmptyStrArray;
+
+                try
+                {
+                    if (!pkgEntry.TryGetProperty(idName, out JsonElement idItem))
+                    {
+                        string errMsg = $"FindNameGlobbing(): Name element could not be found in response.";
+                        edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
+                        return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+                    }
+
+                    if (!pkgEntry.TryGetProperty(tagsName, out JsonElement tagsItem))
+                    {
+                        string errMsg = $"FindNameGlobbing(): Tags element could not be found in response.";
+                        edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
+                        return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+                    }
+
+                    id = idItem.ToString();
+                    pkgTags = GetTagsFromJsonElement(tagsElement: tagsItem);
+
+                    // determine if id matches our wildcard criteria
+                    if ((packageName.StartsWith("*") && packageName.EndsWith("*") && id.ToLower().Contains(querySearchTerm.ToLower())) ||
+                        (packageName.EndsWith("*") && id.StartsWith(querySearchTerm, StringComparison.OrdinalIgnoreCase)) ||
+                        (packageName.StartsWith("*") && id.EndsWith(querySearchTerm, StringComparison.OrdinalIgnoreCase)))
+                    {
+                        bool isTagMatch = DeterminePkgTagsSatisfyRequiredTags(pkgTags: pkgTags, requiredTags: tags);
+                        if (!isTagMatch)
+                        {
+                            continue;
+                        }
+
+                        matchingResponses.Add(pkgEntry.ToString());
+                    }
+                }
+                    
+                catch (Exception e)
+                {
+                    string errMsg = $"FindNameGlobbing(): Name or Version element could not be parsed from response due to exception {e.Message}.";
+                    edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
+                    break;
+                }
+                
+            }
+
+            return new FindResults(stringResponse: matchingResponses.ToArray(), hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+        }
+
+        /// <summary>
+        /// Helper method called by FindTags() for special case where repository is NuGet.org repository.
+        /// </summary>        
+        private FindResults FindTagsFromNuGetRepo(string[] tags, bool includePrerelease, out ExceptionDispatchInfo edi)
+        {
+            Dictionary<string, string> resources = GetResourcesFromServiceIndex(out edi);
+            if (edi != null)
+            {
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            string searchQueryServiceUrl = FindSearchQueryService(resources, out edi);
+            if (edi != null)
+            {
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            string tagsQueryTerm = String.Join(" ", tags);
+
+            string query = $"{searchQueryServiceUrl}?q=tags:{tagsQueryTerm}&prerelease={includePrerelease}&semVerLevel=2.0.0";
+
+            // Get responses for all packages that contain the required tags
+            JsonElement[] tagPkgs = GetJsonElementArr(query, dataName, out edi);
+            if (edi != null)
+            {
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            List<string> matchingResponses = new List<string>();
+            foreach (var pkgEntry in tagPkgs)
+            {
+                try
+                {
+                    matchingResponses.Add(pkgEntry.ToString());
+                }
+                catch (Exception e)
+                {
+                    string errMsg = $"FindTag(): Id or Version element could not be parsed from response due to exception {e.Message}.";
+                    edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
+                    return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+                }
+            }
+
+            return new FindResults(stringResponse: matchingResponses.ToArray(), hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+        }
+
+        /// <summary>
+        /// Helper method called by FindName() and FindNameWithTag()
+        /// <summary>
+        private FindResults FindNameHelper(string packageName, string[] tags, bool includePrerelease, ResourceType type, out ExceptionDispatchInfo edi)
+        {
+            Dictionary<string, string> resources = GetResourcesFromServiceIndex(out edi);
+            if (edi != null)
+            {
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            string registrationsBaseUrl = FindRegistrationsBaseUrlForAllV3(resources, out edi);
+            if (edi != null)
+            {
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            string[] versionedResponses = GetVersionedResponses(registrationsBaseUrl, packageName, catalogEntryProperty, out edi);
+            if (edi != null)
+            {
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            string latestVersionResponse = String.Empty;
+            foreach (string response in versionedResponses)
+            {
+                JsonElement pkgVersionElement;
+                try
+                {
+                    JsonDocument pkgVersionEntry = JsonDocument.Parse(response);
+                    JsonElement rootDom = pkgVersionEntry.RootElement;
+                    if (!rootDom.TryGetProperty(versionName, out pkgVersionElement))
+                    {
+                        edi = ExceptionDispatchInfo.Capture(new InvalidOrEmptyResponse($"Response does not contain '{versionName}' element."));
+                        return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+                    }
+                }
+                catch (Exception e)
+                {
+                    edi = ExceptionDispatchInfo.Capture(e);
+                    return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+                }
+
+                if (NuGetVersion.TryParse(pkgVersionElement.ToString(), out NuGetVersion pkgVersion))
+                {
+                    if (!pkgVersion.IsPrerelease || includePrerelease)
+                    {
+                        // versions are reported in descending order i.e 5.0.0, 3.0.0, 1.0.0 so grabbing the first match suffices
+                        latestVersionResponse = response;
+
+                        // TODO: technically we could call GetTagsFromJsonElement() and DeterminePkgTagsSatisfyRequiredTags() so we don't parse into JsonDocument again
+                        break;
+                    }
+                } 
+            }
+
+            if (String.IsNullOrEmpty(latestVersionResponse))
+            {
+                string errMsg = $"FindName(): Package with Name {packageName} was not found in repository {Repository.Name}.";
+                edi = ExceptionDispatchInfo.Capture(new SpecifiedTagsNotFoundException(errMsg));
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            bool isTagMatch = DetermineTagsPresent(response: latestVersionResponse, tags: tags, out edi);
+            if (!isTagMatch)
+            {
+                if (edi == null)
+                {
+                    string errMsg = $"FindName(): Package with Name {packageName} and Tags {String.Join(", ", tags)} was not found in repository {Repository.Name}.";
+                    edi = ExceptionDispatchInfo.Capture(new SpecifiedTagsNotFoundException(errMsg));
+                }
+
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            return new FindResults(stringResponse: new string[] { latestVersionResponse }, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+        }
+        
+        /// <summary>
+        /// Helper method called by FindVersion() and FindVersionWithTag()
+        /// </summary>
+        private FindResults FindVersionHelper(string packageName, string version, string[] tags, ResourceType type, out ExceptionDispatchInfo edi)
+        {
+            if (!NuGetVersion.TryParse(version, out NuGetVersion requiredVersion))
+            {
+                edi = ExceptionDispatchInfo.Capture(new ArgumentException($"Version {version} to be found is not a valid NuGet version."));
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            // Get all the resources for the service index
+            Dictionary<string, string> resources = GetResourcesFromServiceIndex(out edi);
+            if (edi != null)
+            {
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            string registrationsBaseUrl = FindRegistrationsBaseUrlForAllV3(resources, out edi);
+            if (edi != null)
+            {
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            string[] versionedResponses = GetVersionedResponses(registrationsBaseUrl, packageName, catalogEntryProperty, out edi);
+            if (edi != null)
+            {
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            string latestVersionResponse = String.Empty;
+            foreach (string response in versionedResponses)
+            {
+                // this assumes latest versions are reported first i.e 5.0.0, 3.0.0, 1.0.0
+                JsonElement pkgVersionElement;
+                try
+                {
+                    JsonDocument pkgVersionEntry = JsonDocument.Parse(response);
+                    JsonElement rootDom = pkgVersionEntry.RootElement;
+                    if (!rootDom.TryGetProperty(versionName, out pkgVersionElement))
+                    {
+                        edi = ExceptionDispatchInfo.Capture(new InvalidOrEmptyResponse($"Response does not contain '{versionName}' element for search with Name {packageName} and Version {version}."));
+                        return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+                    }
+                }
+                catch (Exception e)
+                {
+                    edi = ExceptionDispatchInfo.Capture(e);
+                    return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+                }
+
+                if (NuGetVersion.TryParse(pkgVersionElement.ToString(), out NuGetVersion pkgVersion))
+                {
+                    if (pkgVersion == requiredVersion)
+                    {
+                        latestVersionResponse = response;
+                        // TODO: technically we could call GetTagsFromJsonElement() and DeterminePkgTagsSatisfyRequiredTags() so we don't parse into JsonDocument again
+                        // regardless of wheter tags are satisfied we would break - if false that just means we had right version but no tags, so search fails and edi is set but response is empty (handled below)
+                        break;
+                    }
+                }
+            }
+
+            if (String.IsNullOrEmpty(latestVersionResponse))
+            {
+                edi = ExceptionDispatchInfo.Capture(new InvalidOrEmptyResponse($"FindVersion(): Package with Name {packageName}, Version {version} was not found in repository {Repository.Name}"));
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            bool isTagMatch = DetermineTagsPresent(response: latestVersionResponse, tags: tags, out edi);
+            if (!isTagMatch)
+            {
+                if (edi == null)
+                {
+                    string errMsg = $"FindVersion(): Package with Name {packageName}, Version {version} and Tags {String.Join(", ", tags)} was not found in repository {Repository.Name}.";
+                    edi = ExceptionDispatchInfo.Capture(new SpecifiedTagsNotFoundException(errMsg));
+                }
+
+                return new FindResults(stringResponse: Utils.EmptyStrArray, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+            }
+
+            return new FindResults(stringResponse: new string[] { latestVersionResponse }, hashtableResponse: emptyHashResponses, responseType: v3FindResponseType);
+        }
+
+        /// <summary>
         /// Helper method that makes the HTTP request for the V3 server protocol url passed in for find APIs.
         /// </summary>
-        private String HttpRequestCall(string requestUrlV3, out ExceptionDispatchInfo edi)
+        private string HttpRequestCall(string requestUrlV3, out ExceptionDispatchInfo edi)
         {
             edi = null;
             string response = string.Empty;
@@ -864,108 +749,251 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
         }
 
         /// <summary>
-        /// Helper method that makes finds the specified V3 server protocol resources from the service index.
+        /// Finds all resources present in the repository's service index.
         /// </summary>
-        private Hashtable FindResourceType(string[] resourceTypeName, out ExceptionDispatchInfo edi)
+        private Dictionary<string, string> GetResourcesFromServiceIndex(out ExceptionDispatchInfo edi)
         {
-            Hashtable resourceHash = new Hashtable();
-            JsonElement[] resources = GetJsonElementArr($"{Repository.Uri}", resourcesName, out edi);
+            Dictionary<string, string> resources = new Dictionary<string, string>();
+            JsonElement[] resourcesArray = GetJsonElementArr($"{Repository.Uri}", resourcesName, out edi);
             if (edi != null)
             {
-                return resourceHash;
+                return resources;
             }
 
-            foreach (JsonElement resource in resources)
+            foreach (JsonElement resource in resourcesArray)
             {
                 try
                 {
-                    if (resource.TryGetProperty("@type", out JsonElement typeElement) && resourceTypeName.Contains(typeElement.ToString()))
+                    if (!resource.TryGetProperty("@type", out JsonElement typeElement))
                     {
-                        // check if key already present in hastable, as there can be resources with same type but primary/secondary instances
-                        if (!resourceHash.ContainsKey(typeElement.ToString()))
-                        {
-                            if (resource.TryGetProperty("@id", out JsonElement idElement))
-                            {
-                                // add name of the resource and its url
-                                resourceHash.Add(typeElement.ToString(), idElement.ToString());
-                            }
-                            else
-                            {
-                                string errMsg = $"@type element was found but @id element not found in service index '{Repository.Uri}' for {resourceTypeName}.";
-                                edi = ExceptionDispatchInfo.Capture(new V3ResourceNotFoundException(errMsg));
-                                return resourceHash;
-                            }
-                        }
+                        edi = ExceptionDispatchInfo.Capture(new JsonParsingException($"@type element not found for resource in service index for repository {Repository.Name}"));
+                        return new Dictionary<string, string>();
+                    }
+
+                    if (!resource.TryGetProperty("@id", out JsonElement idElement))
+                    {
+                        edi = ExceptionDispatchInfo.Capture(new JsonParsingException($"@id element not found for resource in service index for repository {Repository.Name}"));
+                        return new Dictionary<string, string>();
+                    }
+
+                    if (!resources.ContainsKey(typeElement.ToString()))
+                    {
+                        // Some resources have a primary and secondary entry. The @id value is the same, so we only choose the primary entry.
+                        resources.Add(typeElement.ToString(), idElement.ToString());
                     }
                 }
                 catch (Exception e)
                 {
-                    string errMsg = $"Exception parsing JSON for respository {Repository.Uri} with error: {e.Message}";
+                    string errMsg = $"Exception parsing service index JSON for respository {Repository.Name} with error: {e.Message}";
                     edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
-                    return resourceHash;
-                }
-
-                if (resourceHash.Count == resourceTypeName.Length)
-                {
-                    break;
+                    return new Dictionary<string, string>();
                 }
             }
 
-            foreach (string resourceType in resourceTypeName)
+            return resources;
+        }
+        
+        /// <summary>
+        /// Gets the resource of type "RegistrationBaseUrl" from the repository's resources.
+        /// A repository can have multiple resources of type "RegistrationsBaseUrl" so it finds the best match according to the guideline comment in the method.
+        /// </summary>
+        private string FindRegistrationsBaseUrlForAllV3(Dictionary<string, string> resources, out ExceptionDispatchInfo edi)
+        {
+            edi = null;
+            string registrationsBaseUrl = String.Empty;
+
+            /**
+            If RegistrationsBaseUrl/3.6.0 exists, use RegistrationsBaseUrl/3.6.0
+            Otherwise, if RegistrationsBaseUrl/3.4.0 exists, use RegistrationsBaseUrl/3.4.0
+            Otherwise, if RegistrationsBaseUrl/3.0.0-rc exists, use RegistrationsBaseUrl/3.0.0-rc
+            Otherwise, if RegistrationsBaseUrl/3.0.0-beta exists, use RegistrationsBaseUrl/3.0.0-beta
+            Otherwise, if RegistrationsBaseUrl exists, use RegistrationsBaseUrl
+            Otherwise, report an error
+            */
+
+            if (resources.ContainsKey("RegistrationsBaseUrl/3.6.0"))
             {
-                if (!resourceHash.ContainsKey(resourceType))
-                {
-                    string errMsg = $"FindResourceType(): Could not find resource type {resourceType} from the service index.";
-                    edi = ExceptionDispatchInfo.Capture(new V3ResourceNotFoundException(errMsg));
-                    break;
-                }
+                registrationsBaseUrl = resources["RegistrationsBaseUrl/3.6.0"];
+            }
+            else if (resources.ContainsKey("RegistrationsBaseUrl/3.4.0"))
+            {
+                registrationsBaseUrl = resources["RegistrationsBaseUrl/3.4.0"];
+            }
+            else if (resources.ContainsKey("RegistrationsBaseUrl/3.0.0-rc"))
+            {
+                registrationsBaseUrl = resources["RegistrationsBaseUrl/3.0.0-rc"];
+            }
+            else if (resources.ContainsKey("RegistrationsBaseUrl/3.0.0-beta"))
+            {
+                registrationsBaseUrl = resources["RegistrationsBaseUrl/3.0.0-beta"];
+            }
+            else if (resources.ContainsKey("RegistrationsBaseUrl"))
+            {
+                registrationsBaseUrl = resources["RegistrationsBaseUrl"];
+            }
+            else
+            {
+                edi = ExceptionDispatchInfo.Capture(new V3ResourceNotFoundException($"RegistrationBaseUrl resource could not be found for Repository '{Repository.Name}'"));
             }
 
-            return resourceHash;
+            return registrationsBaseUrl;
         }
 
         /// <summary>
-        /// Helper method finds package with name and specified version
-        /// <summary>
-        private string FindVersionHelper(string registrationsBaseUrl, string packageName, string version, out ExceptionDispatchInfo edi)
+        /// Gets the resource of type "SearchQueryService" from the repository's resources.
+        /// A repository can have multiple resources of type "SearchQueryService" so it finds the best match according to the guideline comment in the method.
+        /// </summary>
+        private string FindSearchQueryService(Dictionary<string, string> resources, out ExceptionDispatchInfo edi)
         {
-            // https://api.nuget.org/v3/registration5-gz-semver2/newtonsoft.json/13.0.2.json
-            var requestPkgMapping = $"{registrationsBaseUrl}{packageName.ToLower()}/{version}.json";
+            edi = null;
+            string searchQueryServiceUrl = String.Empty;
+
+            if (resources.ContainsKey("SearchQueryService/3.5.0"))
+            {
+                searchQueryServiceUrl = resources["SearchQueryService/3.5.0"];
+            }
+            else if (resources.ContainsKey("SearchQueryService/3.0.0-rc"))
+            {
+                searchQueryServiceUrl = resources["SearchQueryService/3.0.0-rc"];
+            }
+            else if (resources.ContainsKey("SearchQueryService/3.0.0-beta"))
+            {
+                searchQueryServiceUrl = resources["SearchQueryService/3.0.0-beta"];
+            }
+            else if (resources.ContainsKey("SearchQueryService"))
+            {
+                searchQueryServiceUrl = resources["SearchQueryService"];
+            }
+            else
+            {
+                edi = ExceptionDispatchInfo.Capture(new V3ResourceNotFoundException($"SearchQueryService resource could not be found for Repository '{Repository.Name}'"));
+            }
+
+            return searchQueryServiceUrl;
+        }
+
+        /// <summary>
+        /// Helper method iterates through the entries in the registrationsUrl for a specific package and all its versions.
+        /// This contains an inner items element (containing the package metadata) and the packageContent element (containing URI through which the .nupkg can be downloaded)
+        /// <param name="property"> This can be the "catalogEntry" or "packageContent" property.
+        ///     The "catalogEntry" property is used for search, and the value is package metadata.
+        ///     The "packageContent" property is used for download, and the value is a URI for the .nupkg file.
+        /// </param>
+        /// <summary>
+        private string[] GetVersionedResponses(string registrationsBaseUrl, string packageName, string property, out ExceptionDispatchInfo edi)
+        {
+            List<string> versionedResponses = new List<string>();
+            string[] versionedResponseArr;
+            var requestPkgMapping = $"{registrationsBaseUrl}{packageName.ToLower()}/index.json";
             string pkgMappingResponse = HttpRequestCall(requestPkgMapping, out edi);
             if (edi != null)
             {
-                return String.Empty;
+                return Utils.EmptyStrArray;
             }
 
-            string catalogEntryUrl = string.Empty;
             try
             {
-                JsonDocument pkgMappingDom = JsonDocument.Parse(pkgMappingResponse);
-                JsonElement rootPkgMappingDom = pkgMappingDom.RootElement;
+                // parse out JSON response we get from RegistrationsUrl
+                JsonDocument pkgVersionEntry = JsonDocument.Parse(pkgMappingResponse);
 
-                if (!rootPkgMappingDom.TryGetProperty("catalogEntry", out JsonElement catalogEntryUrlElement))
+                // The response has a "items" array element, which only has useful 1st element
+                JsonElement rootDom = pkgVersionEntry.RootElement;
+                rootDom.TryGetProperty(itemsName, out JsonElement itemsElement);
+                if (itemsElement.GetArrayLength() == 0)
                 {
-                    string errMsg = $"FindVersionHelper(): CatalogEntry element could not be found in response or was empty.";
-                    edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
-                    return String.Empty;
+                    edi = ExceptionDispatchInfo.Capture(new ArgumentException($"Response does not contain '{itemsName}' element, for package with Name {packageName}."));
+                    return Utils.EmptyStrArray;
                 }
 
-                catalogEntryUrl = catalogEntryUrlElement.ToString();
+                JsonElement firstItem = itemsElement[0];
+
+                // The "items" property contains an inner "items" element and a "count" element
+                // The inner "items" property is the metadata array for each version of the package.
+                // The "count" property represents how many versions are present for that package, (i.e how many elements are in the inner "items" array)
+                // JsonElement innerItemsElements = firstItem.GetProperty(itemsName);
+                if (!firstItem.TryGetProperty(itemsName, out JsonElement innerItemsElements))
+                {
+                    edi = ExceptionDispatchInfo.Capture(new ArgumentException($"Response does not contain inner '{itemsName}' element, for package with Name {packageName}."));
+                    return Utils.EmptyStrArray;
+                }
+                
+                if (!firstItem.TryGetProperty(countName, out JsonElement countElement) || !countElement.TryGetInt32(out int count))
+                {
+                    edi = ExceptionDispatchInfo.Capture(new ArgumentException($"Response does not contain inner '{countName}' element or it is not a valid integer, for package with Name {packageName}."));
+                    return Utils.EmptyStrArray;
+                }
+
+                if (!firstItem.TryGetProperty("upper", out JsonElement upperVersionElement))
+                {
+                    edi = ExceptionDispatchInfo.Capture(new ArgumentException($"Response does not contain inner 'upper' element, for package with Name {packageName}."));
+                    return Utils.EmptyStrArray;
+                }
+
+                for (int i = 0; i < count; i++)
+                {
+                    // Get the specific entry for each package version
+                    JsonElement versionedItem = innerItemsElements[i];
+                    if (!versionedItem.TryGetProperty(property, out JsonElement metadataElement))
+                    {
+                        edi = ExceptionDispatchInfo.Capture(new ArgumentException($"Response does not contain inner '{property}' element, for package with Name {packageName}."));
+                        return Utils.EmptyStrArray;
+                    }
+                    
+                    versionedResponses.Add(metadataElement.ToString());
+                }
+
+                // Reverse array of versioned responses, if needed, so that version entries are in descending order.
+                string upperVersion = upperVersionElement.ToString();
+                versionedResponseArr = versionedResponses.ToArray();
+                if (!IsLatestVersionFirst(versionedResponseArr, upperVersion, out edi))
+                {
+                    Array.Reverse(versionedResponseArr);
+                }
+
             }
             catch (Exception e)
             {
-                string errMsg = $"FindVersionHelper(): Exception parsing JSON for respository {Repository.Uri} with error: {e.Message}";
-                edi = ExceptionDispatchInfo.Capture(new JsonParsingException(errMsg));
-                return String.Empty;
+                edi = ExceptionDispatchInfo.Capture(e);
+                return Utils.EmptyStrArray;
             }
 
-            string response = HttpRequestCall(catalogEntryUrl, out edi);
-            if (edi != null)
+            return versionedResponseArr;
+        }
+
+        /// <summary>
+        /// Returns true if the metadata entries are arranged in descending order with respect to the package's version.
+        /// ADO feeds usually return version entries in descending order, but Nuget.org repository returns them in ascending order.
+        /// </summary>
+        private bool IsLatestVersionFirst(string[] versionedResponses, string upperVersion, out ExceptionDispatchInfo edi)
+        {
+            edi = null;
+            bool latestVersionFirst = true;
+
+            // We don't need to perform this check if no responses, or single response
+            if (versionedResponses.Length < 2)
             {
-                return String.Empty;
+                return latestVersionFirst;
             }
 
-            return response;
+            string firstResponse = versionedResponses[0];
+            JsonDocument firstResponseJson = JsonDocument.Parse(firstResponse);
+            JsonElement firstResponseDom = firstResponseJson.RootElement;
+            if (!firstResponseDom.TryGetProperty(versionName, out JsonElement firstVersionElement))
+            {
+                edi = ExceptionDispatchInfo.Capture(new JsonParsingException($"Response did not contain '{versionName}' element"));
+                return latestVersionFirst;
+            }
+
+            string firstVersion = firstVersionElement.ToString();
+            if (NuGetVersion.TryParse(upperVersion, out NuGetVersion upperPkgVersion) && NuGetVersion.TryParse(firstVersion, out NuGetVersion firstPkgVersion))
+            {
+                if (firstPkgVersion != upperPkgVersion)
+                {
+                    latestVersionFirst = false;
+                }
+            }
+
+            return latestVersionFirst;
         }
 
         /// <summary>
@@ -1037,31 +1065,11 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
         }
 
         /// <summary>
-        /// Helper method that returns a flattened list of all versions present for a package.
-        /// Implementation note: NuGet server and Azure Artifacts server return this flattened version list in opposite orders so we reverse array accordingly.
-        /// </summary>
-        private JsonElement[] GetPackageVersions(string packageBaseAddressUrl, string packageName, bool isNuGetRepo, out ExceptionDispatchInfo edi)
-        {
-            if (String.IsNullOrEmpty(packageBaseAddressUrl))
-            {
-                edi = ExceptionDispatchInfo.Capture(new ArgumentException($"GetPackageVersions(): Package Base URL cannot be null or empty"));
-                return new JsonElement[]{};
-            }
-
-            JsonElement[] pkgVersionsElement = GetJsonElementArr($"{packageBaseAddressUrl}{packageName.ToLower()}/index.json", versionsName, out edi);
-            if (edi != null)
-            {
-                return new JsonElement[]{};
-            }
-
-            return isNuGetRepo ? pkgVersionsElement.Reverse().ToArray() : pkgVersionsElement.ToArray();
-        }
-
-        /// <summary>
         /// Helper method that parses response for given property and returns result for that property as a JsonElement array.
         /// </summary>
         private JsonElement[] GetJsonElementArr(string request, string propertyName, out ExceptionDispatchInfo edi)
         {
+            // TODO: for "data" propertyName, basically SearchQueryService, if "totalHits" == 0 no matches were found, so populate edi and caller and early out.
             JsonElement[] pkgsArr = new JsonElement[0];
             try
             { 
@@ -1089,7 +1097,7 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
         /// <summary>
         /// Helper method called by HttpRequestCall() that makes the HTTP request for string response.
         /// </summary>
-        public static async Task<string> SendV3RequestAsync(HttpRequestMessage message, HttpClient s_client)
+        private static async Task<string> SendV3RequestAsync(HttpRequestMessage message, HttpClient s_client)
         {
             string errMsg = "SendV3RequestAsync(): Error occured while trying to retrieve response: ";
 
@@ -1119,7 +1127,7 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
         /// <summary>
         /// Helper method called by HttpRequestCallForContent() that makes the HTTP request for string response.
         /// </summary>
-        public static async Task<HttpContent> SendV3RequestForContentAsync(HttpRequestMessage message, HttpClient s_client)
+        private static async Task<HttpContent> SendV3RequestForContentAsync(HttpRequestMessage message, HttpClient s_client)
         {
             string errMsg = "SendV3RequestForContentAsync(): Error occured while trying to retrieve response for content: ";
